requirements.txt
openai-whisper==20240930
pydub==0.25.1